[REFACTOR] 패키지 분리
