+++ operation-api/src/main/java/org/operation/app/attendance/api/AttendanceApi.java
@@ -1,4 +1,4 @@
-package org.operation.app.attendance;
+package org.operation.app.attendance.api;
 
 import java.security.Principal;
